+++ src/components/logIn/InputPhone.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useState } from "react";
-import { InputField } from "@/components/inputField";
+import { InputField } from "@/src/components/inputField";
 
 import { SignUpFormData } from "@/type/authData";
